+++ packages/orq-rc/RELEASES.md
@@ -10589,3 +10589,13 @@ Based on:
 - [python v4.15.0-rc.11] packages/orq-rc
 ### Releases
 - [PyPI v4.15.0-rc.11] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.11 - packages/orq-rc
+
+## 2026-09-02 06:44:22
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.2 (2.933.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.15.0-rc.12] packages/orq-rc
+### Releases
+- [PyPI v4.15.0-rc.12] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.12 - packages/orq-rc